solution

+++ restaurants/views.py
@@ -59,6 +59,9 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +78,8 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or request.user == restaurant.owner):
+        return redirect('access_denied')
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -83,14 +88,17 @@ def item_create(request, restaurant_id):
             item.save()
             return redirect('restaurant-detail', restaurant_id)
     context = {
-        "form":form,
-        "restaurant": restaurant,
-    }
+            "form":form,
+            "restaurant": restaurant,
+        }
     return render(request, 'item_create.html', context)
+
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     form = RestaurantForm(instance=restaurant_obj)
+    if not (request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect('access_denied')
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
         if form.is_valid():
@@ -99,10 +107,17 @@ def restaurant_update(request, restaurant_id):
     context = {
         "restaurant_obj": restaurant_obj,
         "form":form,
-    }
+        }
     return render(request, 'update.html', context)
 
+
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect('access_denied')
+
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def access_denied(request):
+    return render(request, 'access_denied.html')
